news/requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
lxml>=5.0.0
pyyaml>=6.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Async Processing
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Logging
loguru>=0.7.0

